Changed to separate header file
